+++ services/backend/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @frak-labs/backend-elysia
 
+## 0.0.24
+
+### Patch Changes
+
+- Updated dependencies [[`9a1798c`](https://github.com/frak-id/wallet/commit/9a1798cfe397f0b6b4dc4049f5d8bab0587cc4d7)]:
+  - @frak-labs/core-sdk@1.1.4
+
 ## 0.0.23
 
 ### Patch Changes